Move _ApplicationSettings to models.ApplicationSettings
Make ApplicationSettings public 'cos it's going to be used in multiple
places soon.
