+++ lms/models/__init__.py
@@ -1,4 +1,5 @@
 from lms.models.application_instance import ApplicationInstance
+from lms.models.application_settings import ApplicationSettings
 from lms.models.grading_info import GradingInfo
 from lms.models.group_info import GroupInfo
 from lms.models.h_group import HGroup, h_group_name
